Improve diagram design
